requirements_test.txt
# Testing requirements for Anode Battery integration
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-homeassistant-custom-component==0.13.106
